레디스 캐시를 적용하여 카테고리 조회 로직 구현

+++ src/main/java/com/flab/makedel/controller/StoreController.java
@@ -82,7 +82,7 @@ public class StoreController {
 
     }
 
-    private void validateMyStore(int storeId, String ownerId) throws HttpClientErrorException {
+    private void validateMyStore(long storeId, String ownerId) throws HttpClientErrorException {
         boolean isMyStore = storeService.isMyStore(storeId, ownerId);
         if (!isMyStore) {
             throw new HttpClientErrorException(HttpStatus.UNAUTHORIZED);
